Code completion now working for C# files in ASP.NET projects.
Change DnxProject to be a DotNetAssembly project to get code
completion to work.

Hard coded the DnxProject language to be C# to fix null ref
on loading the project.

Update the assembly references in the solution window after the
information is received from dnx. Currently only showing the
references for the first framework that the project supports.
SharpDevelop has no support for changing the current code
completion context in the text editor.