Fix invalid configuration mapping message in Solution window.
Opening a ASP.NET 5 console project would show an invalid
configuration mapping message in the Solution window for the console
project. This was because a matching configuration could not be found
for the Debug configuration in the project. This was because the
configurations created for the DNX project were using the full name
"Debug|Any CPU" when it should have been just "Debug".