Prevent .csproj file being created when saving .xproj
Saving is currently disabled for .xproj files. Previously saving the
project would generate a .csproj file.

Previously opening an .xproj directly with no solution file .sln in
the same directory would generate a solution file pointing to a
.csproj file.

The .csproj file extension was being used since the
SolutionEntityItem's FileName property when being set would call the
FileFormat's GetValidFileName method. This method would find the
DotNetProjectNode for csproj files when looking at the
/MonoDevelop/ProjectModel/MSBuildItemTypes extension. This would then
cause the project filename extension to be changed to .csproj. There
does not seem to be a way to insert the DotNetProjectNode for
DnxProjects since these extensions do not have an id. So for now
the DnxProject overrides the SolutionEntityItem's FileName property
and does not call FileFormat.GetValidFileName().