Add custom project template generator.
This is currently just using the existing MonoDevelop project
template generator but it is hosted in the addin. The code has been
copied from MonoDevelop to the addin. There is no added features in
the template generator. This is just the first stage of getting it
working but unchanged from what is in MonoDevelop.

Also having to workaround a bug in MonoDevelop where the project
category is removed before the custom project template provider
can add its project template. So the ASP.NET 5 console project
template was not being displayed in the New Project dialog. The
workaround is to have a dummy project template with the same
metadata, with a group id, so it is grouped with the other console
project template but with a condition so it is never selected. This
dummy project template is registered with MonoDevelop's existing
project template provider but never used.