Use project template wizard to generate project instead of template.
Added a library project that uses the project template wizard to add
solution folders and the library project to the solution instead of
using a custom project template provider.

Configurations not being created in solution.
Also .xproj file is being saved with the wrong MSBuildProjectHandler
so the content of the project file look like a C# project file. If
the other DnxProject constructor is used that does not pass the
language to the DotNetAssemblyProject base class then the Solution
Pad has an error next to the project since the language is not set.