Fix .xproj file having incorrect content after saving.
When the .xproj file was saved by the project template wizard
the wrong MSBuildProjectHandler was being used.

In order to get the DnxMSBuildProjectHandler to be used a hack has
been implemented. A lot of the methods required to replace the
ItemHandler for the project are internal. So the only way to register
a new item handler and setup the EntityItem for the handler is to
create a class that derives from ItemTypeNode and override its
OnCreateHandler method to return the correct handler.