DNX runtime version can be changed in project options.
Added a project options panel: Build - DNX Runtime which allows the
DNX SDK version number to be changed in the global.json file.

This is a project options panel and not a solution options panel so
option can be hidden in the dialog if the project is not a DNX
project. Custom conditions to show/hide options for a solution are
not possible. This would mean the DNX runtime options page could not
be hidden if a solution did not contain any DNX projects.

The list of DNX runtime versions is currently hard coded. The
installed versions should be determined and listed instead.

The format of the global.json file is currently fixed to use tabs.
It should use a custom code formatter to allow the formatting to be
customised.