Fix being unable to run new project.
When a new solution was created the main toolbar did not show any
execution targets taken from the project.json file. This was because
the project was added to the solution by the wizard after it was
opened and the solution had no startup project set. Without a startup
project any changes to the execution targets were not reflected in
the main toolbar. This then resulted in no command name being used
when running the application so the dnx help output was shown in the
console window.