New web project template uses Kestrel when not on Windows.
The original web project template used IIS and the WebListener which
are only available on Windows. Running this web project on the Mac
would fail with show an exception in the output from dnx.

Added a separate project template with associated files to use
Kestrel. Ideally this would have used conditional code in the
file templates but unfortunately it is not possible to pass parameters
directly to the FileDescriptionTemplate since the required method
is internal. So instead there is another project template which is
used when not on Windows.