Project references now reflect the selected execution target.
Changing the execution target will cause the Solution window to update
so the References node for the project shows the references for the
selected DNX framework.

If the default execution target profile is selected then the first
DNX framework is used as returned from the DNX host. This is probably
incorrect.

Also using the execution target to change the references is not the
best approach. Ideally the path selector in the text editor would be
used, in the same way it works in Visual Studio, so the DNX framework
selected there would determine the references being used.
Unfortunately this would require a change to MonoDevelop so for now
the execution target is being used.

Type information is updated when the execution target is changed since
the references are updated. Unfortunately types in mscorlib are not
resolved correctly so DateTime.Now.ToShortDateString () when using
DNX Core shows no errors when it should since it does not exist for
DNX Core.