Fix diagnostic message pattern matching on Mac/Linux.
The diagnostic messages returned from the DNX host when a build
is triggered in the IDE were not being parsed correctly on the Mac
and only worked on Windows. This meant the filenames for the
errors and warnings was not parsed so it was not possible to click
them in the Errors window to jump to the corresponding line in the
text editor.