Support adding NuGet package references to DNX projects.
The Dependencies folder in the Solution pad now has an Add NuGet
packages menu. This opens the Add Packages dialog. NuGet packages
selected will be added to the dependencies in the project.json file.

There is no handling of NuGet packages that are not compatible with
DNX projects. For example, if AvalonEdit is installed into Visual
Studio 2015 then the NuGet package is only added as a dependency into
the frameworks/dnx451 section of the project.json file.