Handle the same NuGet package being installed.
Previously it was not possible to install a different version of a
NuGet package when there was already a dependency in the project.json
file. This was because a duplicate NuGet package dependency was being
added to the project.json file and failing since duplicate properties
are not allowed. Now a check is made for an existing dependency with
the same name and it is replaced.