Fix output missing from DNX Output window.
Ensure that output is available to be displayed in the DNX Output
window even if the pad is not created.

Fix DNX output being cleared when the solution is loaded. This was
causing some of the output generated as the solution is being loaded
to be cleared from the DNX Output.