Improve error handling when trying to start the DNX host.
Do not show an error dialog when an exception is thrown when trying
to create the DNX project system. This can happen currently if
MONO_MANAGED_WATCHER is not set to false and the FileSystemWatcher
hits a limit of the number of files to watch.

Handle the DNX runtime not being found. Previously when a build was
run you would see:

   Build failed. Value cannot be null

With no obvious way to fix the problem. Now the error about being
unable to find the DNX runtime will be shown in the build error
message.

In both the above cases if a build is run a more useful error will
be displayed as the build error. Also the Dependencies folder will
show an error icon with a tooltip that shows the error.