Fix code completion with latest monodevelop master branch.
Project.GetSourceFilesAsync () has changed in monodevelop so that
instead of taking the Files from the project instead the files are
taken MSBuild when it evaluates the items. This happens even though
the DNX project has UseMSBuildEngine set to false. For DNX projects
(.xproj) files there are no files defined in the MSBuild file so the
type system parses no files and code completion was not working.

To workaround this the files are added to the MSBuildProject
associated with the DotNetProject so the TypeSystem has files to
parse and code completion works.

Fixed #7