Improve error reporting when failing to find DNX runtime.
If the runtime was missing when creating a new project an exception
was thrown and the error message tooltip on the dependencies folder
in the Solution window was empty.

When the global.json file is changed so it has an invalid runtime the
dependencies folder node in the Solution window was not being updated
to show an error.