Fix ToolsVersion in DNX projects created by Xamarin Studio 5.10
Use ToolsVersion="14.0" in Xamarin Studio 5.10. With the existing
implementation the ToolsVersion was always set to 4.0

The change made in commit 90d8f843ce0e5bbd0faf6e91fbeda5906e4ee13e
to make the .xproj files compatible with Xamarin Studio 5.9 did not
set the correct ToolsVersion to 14.0 for projects created by Xamarin
Studio 5.10. The problem is that the constants defined on the
BuildInfo are not static readonly strings but const strings so the
compiler replaces the BuildInfo field references with string constants
so it is not possible to get the current IDE version without using
reflection to force the BuildInfo field to be used for the current
IDE.