Watch files in directories defined in global.json file.
Previously all files were watched inside the solution directory. This
can cause the "kqueue() FileSystemWatcher has reached the maximum
number of files to watch" error on Mono if the solution directory also
includes files for version control, such as a .git subdirectory. So
now the global.json file is read and only the subdirectories defined
in that file are monitored.

Fixes #15