Fix being unable to create new DNX solution.
A null reference was thrown by the main toolbar when a new solution
was created from the New Project dialog. The New Project dialog
would not close and the project was not fully created.

Setting the startup project for the solution before adding the
configurations to the solution works around the following
MonoDevelop bug:

https://bugzilla.xamarin.com/show_bug.cgi?id=38338

Fixes issue #20