Fix code completion not working.
A change was made to MonoDevelop that excluded projects from the type
system which stopped code completion working for DNX projects since
they did not have a C# or VB.NET project type guid. MonoDevelop has
been changed so that the DotNetProject has a SupportsRoslyn flag which
can be set to true in a project so that it is loaded by the type
system. This is set to true for DNX projects so code completion now
works again.

Fixes #18