Fix file items being added to .xproj
When a new project was created the .xproj would be saved and it would
contain the files added to the project. Reloading this new project
would result in duplicate files shown in the Solution window since
they were added twice.

To fix this the project extension no longer relies on overriding
OnWriteProject since the MSBuildProject given to this method has the
files added to it in order so the type system provides code completion
for the project. Now instead the project extension completely takes
over the saving of the project file by overriding OnSave and a new
MSBuildProject is created, without the files added for the type system
and this is then saved.