Fix file not being opened when creating new project.
The change in the previous commit seems to have caused a problem where
when the solution is saved the project's ParentSolution is set to
null. The first time a new DNX solution is created the ParentSolution
is not null and the file is opened into the text editor. The next time
a solution is created or a new project is added to the existing
solution as soon as the solution is saved the ParentSolution for the
project gets set to null for some reason. For now I have just worked
around this by opening the file before the solution is saved since at
this point in time the project's ParentSolution is not null.