Do not start restore if project.lock.json file changes.
The current restore logic is taken from OmniSharp. The latest
OmniSharp code no longer triggers a restore when the project.lock.json
file is changed.

This may or may not reduce the number of times a restore is called
which was reported in issue #23

Occasionally multiple restore processes are started for a single
project when the project.json file is changed. This seems to be
caused by the lock file being changed since there are many (40+)
file watcher events triggered when a restore is run. Now that the
lock file being changed no longer triggers a restore then hopefully
multiple restores for the same underlying project.json file change
should not occur.