Fix null reference when creating new DNX solution.
When creating a new DNX solution a null reference would be logged:

System.NullReferenceException: Object reference not set to an instance of an object
  at MonoDevelop.Projects.Project.OnDispose () in monodevelop/main/src/core/MonoDevelop.Core/MonoDevelop.Projects/Project.cs:877
  at MonoDevelop.Projects.DotNetProject.OnDispose () in monodevelop/main/src/core/MonoDevelop.Core/MonoDevelop.Projects/DotNetProject.cs:531
  at MonoDevelop.Projects.WorkspaceObject.Dispose () in monodevelop/main/src/core/MonoDevelop.Core/MonoDevelop.Projects/WorkspaceObject.cs:223
  at MonoDevelop.Projects.SolutionFolder+<ReloadItem>c__async0.MoveNext () in monodevelop/main/src/core/MonoDevelop.Core/MonoDevelop.Projects/SolutionFolder.cs:258

When this was logged the References in the Solution window would be
empty and no type information was available so the text editor would
show errors for any C# code. The underlying problem was that the
DNX addin overrides the default save behaviour and so the
FileStatusWatcher is not updated and so the project file is reloaded
during the new project creation process. This then causes the
project to lose any reference information it had. Now when the
.xproj file is saved the NeedsReload is set to false on the project.

Unfortunately fixing the reload causes the null reference exception
to now occur when closing the solution. This results in an error
dialog being shown. This only happens when a new DNX solution is
created and then closed, not when loading an existing solution and
then closing it. The problem here is that the Project's OnDispose
method expects the internal ProjectInstance property to be set. This
is not set when creating a new DNX project but is when the project is
loaded from disk. Since the ProjectInstance is internal a workaround
has been implemented that sets the ProjectInstance if it is null by
using reflection. Ideally the Project's OnDispose method would handle
the ProjectInstance being null, it already handles this when a
configuration is removed.