Show diagnostic errors/warnings in Errors window.
Errors and warnings about the project.json file are now displayed in
the Errors window when they are received from the design time host.
Using the DependencyDiagnostics message instead of the Diagnostics
message since the Diagnostics message seems to have a bug where it
keeps appending the same messages to its list. The
DependencyDiagnostics message does not do this but sends a message
for each supported target framework in the project so they are
currently filtered based on the currently selected active framework.