Try to get the output path from .NET Core.
Using the Microsoft.DotNet.ProjectModel to try to find the
output executable or dll for a .NET Core project.

Unfortunately the project model assumes that the current process
has the correct architecture, since normally the application would
be run via dotnet. This causes a problem since we are using the
project model in MonoDevelop so the architecture (x86 or x64) is
returned based on the MonoDevelop process and not the dotnet process
so the wrong executable path is returned for projects that target
the full .NET framework and not .NET Core.