Do not check alternative architecture for executable.
On Windows the debugger can only debug applications for the same
architecture as the host IDE process so in this case the addin falls
back to just using 'dotnet run' instead of looking for alternative
architecture for the compiled executable. If this is not done then
an error message is displayed - 'Debugger operation failed. The
request is not supported (0x80070032)'. Which is the same error you
see when trying to debug a console project that has been compiled for
x64.

Currently only Windows supports both x86 and x64 versions of .NET Core
dotnet executable so making this change for all operating systems
should be OK.