Debugging on Windows now working for projects targeting full .NET
You can now create a .NET Core console project, change the framework
in the project.json so it only targets net451 and then build and
debug it if you have the x86 version of the .NET Core SDK installed.
If you have the x64 version of .NET Core SDK then the addin will
fallback to using 'dotnet run' and there will be no debugger support.