Try other architecture for output file when not on Windows.
When using Mono it is possible to debug a x64 executable even though
the current IDE process is x86. So when looking for the output
executable the addin now tries both the current IDE architecture and
then the alternative architecture if no file is found.

This basically reverts the commit below but adds the restriction to
only check other architectures when not on Windows:

55a88e83416de080c14467f48e2fa6cee3eb12d0