Support running against a particular target framework.
The target frameworks supported by the project are shown as execution
targets in the main toolbar. Selecting a particular framework will
run or debug the application for that framework. Previously commands
were shown as execution targets, with each command available for
each of the target frameworks. Now commands are not shown in the
main toolbar.