Handle framework being removed from project.json
When a framework was removed from the project.json it was still cached
with the project. This would cause the dependencies folder to still
show the old framework.

Now when a framework is removed all cached information relating to the
old framework is removed. Also the current framework being used for the
project is cleared if it was the one removed. Clearing the current
framework makes sure the references and preprocessor information is
updated.