Fix build errors using .NET 4.5.1.
No longer using the addin maker. This fixes being unable to build the
projects since they target .NET 4.5.1 and the addin maker does not
support loading projects that target anything higher than .NET 4.5.

Building now assumes that the MonoDevelop binaries are available in
a monodevelop\main\build\bin directory which will be the case if
the MonoDevelop source code is cloned into a monodevelop directory
in the same parent directory used for the addin's source code. The
location can also be overridden by using the MDBinDir property.