Fix being unable to open a project created on Windows.
With the .NET Core custom MSBuild targets installed a .NET Core
project could be created but then not opened again. The IDE
would show a message saying that the project does not support
.NETFramework,Version=v4.5,Profile=Client. The Microsoft.DotNet.Props
file sets the TargetFrameworkProfile to Client:

  <TargetFrameworkProfile>Client</TargetFrameworkProfile>

The IDE marks the client profile as only supporting the .NET 4.0 CLR
and the project fails to load since the XProject explicitly states it
only supports .NET 4.5 or above since .NET Core projects only support
.NET 4.5 or above. To fix this a check is made in the overridden
OnSupportsFramework method so that if the profile is Client and
the .NET Framework is 4.5 or higher the XProject indicates that this
framework is supported which allows the project to be loaded.

Fixes #28