Support projects created by VS 2015 Update 2.
Visual Studio 2015 Update 2 creates .xproj files with a ToolsVersion
set to 14.0.25123.0 which is not currently supported by Xamarin
Studio. This means the .xproj files fail to load with the
warning 'Unknown solution item type' displayed.

Added a workaround for this by using a custom WorkspaceObjectReader
that will load any .xproj file that it finds independent of the
ToolsVersion setting. This will not be used if the ToolsVersion is
supported by Xamarin Studio since it is the last extension to be
checked.

Fixes #32