Added initial support for running unit tests.
If a testRunner has been added to the project.json file then the
project will be displayed in the Unit Tests window. From the Unit
Test window it is currently possible to run all unit tests and
see the output from dotnet test in the Application Output and a
summary in the Test Results window. Individual tests are not shown
in the Unit Test window nor in the Test Results window currently.

The addin is using the dotnet test protocol to do this rather than
just running 'dotnet test'.

https://dotnet.github.io/docs/core-concepts/core-sdk/cli/dotnet-test-protocol.html

The addin opens a listening socket then runs

    dotnet test --port Number --no-build

Then waits for dotnet test to connect back. Then the addin requests
the process start information for the test runner and launches it.
Test results are then sent back to the listening socket and processed.