Added initial support for test discovery.
Currently the tests are only discovered the first time the Unit Tests
window is opened. The tests are discovered using the dotnet test
protocol. A server is started listening on a port and then dotnet test
is run specifying this port. Once dotnet test has connected back to
the port and sent a test connected message the addin then sends a
start discovery message to the connected client and the test
information is then sent back to the addin from the test runner.