Test runner now updates some of the test information
Test result information is now shown for each test in the Test Results
window. Test results are now shown in the Unit Tests pad for test
methods.

Namespaces in the Unit Tests window are not updated correctly so they
are not marked in green or red when the tests finish.

When there is a slow running test the Unit Tests window does not show
the currently running test.

Still not possible to run individual tests or all tests within a
particular namespace.

Added a workaround for Json.NET not correctly deserializing a nullable
LineNumber. For test discovery if the line number is nullable in the
Test class then the line number was incorrect. So for test discovery a
non-nullable integer is used. This means the same Test class cannot
be used when reading the test results since this returns a null value
for the LineNumber.