Filter console output from design time host based on the log level
The console output from the design time host shows all the json
messages so the console output shown in the .NET Core output window is
now filtered based on the current log level set in preferences and
only shown if the logging is set to Verbose. The console output when
running dotnet test to discovery tests is also filtered bassed on the
current log level.