Fix package restore happening continually.
If a dependency cannot be resolved then package restore would
repeatedly be triggered. The original OmniSharp code would start a
package restore if dependencies were unresolved. Then after the
restore process exits a RestoreCompleted message was sent to the
design time host. The design time host then sends another message
back indicating that dependencies are not resolved. This again
triggers the package restore to be run.

The DnxProjectSystem has been changed so it does not try to
restore if there are missing dependencies. Now the DnxProject itself
will trigger restore if there are missing dependencies only the once
after a project has been opened. This prevents the restore from being
triggered repeatedly due to missing dependencies.

#30 (Issue 2, 6 and 8)