Add info to addin description about uninstalling old addin
When the new version of the addin is installed without the old
version being uninstalled an error about the test provider class being
missing is reported. The addin will then not properly load.
Uninstalling the old addin, restarting the IDE, then installing the
addin works around this problem. The addin description now says that
the old addin must be uninstalled first and the IDE restarted before
the new version is installed.