Stop test discovery if no tests found after a period of time
If there is no assembly compiled and a testRunner is defined then
the dotnet test loader will never finish and does not return any
errors. This would leave the Unit Tests window always showing that
it was loading the tests but it would never complete. If the project
was then compiled the Unit Tests window would not update.

Now there is a timer that will kill the dotnet test process if no
tests have been discovered after 10 seconds. The problem is reported
in the .NET Core Output window and the Unit Tests window will show
that the load failed. When the project is next built after this
shutdown timer has fired then tests should be displayed in the Unit
Tests window.