Handle testRunner being added to project.json
Previously when a testRunner was added to a project.json file and
then the project was compiled the project would not be shown in the
Unit Tests window until the solution was closed and reopened again.

The normal approach to getting projects to appear in the Unit Tests
window is when the project is changed, such as when a reference is
added or removed, the UnitTestService's RebuildTests method is
called which allows the test providers to refresh the list of
available projects to test. Unfortunately the RebuildTests method
is a private method and the .NET Core addin does not add or remove
references. So without resorting to reflection there is no nice
way of triggering this rebuild of the test information. So instead the
addin always adds every .NET Core project to the Unit Tests window.
If a project has no testRunner the Unit Tests window shows this
information. The Unit Tests window will be updated after a build so
if a testRunner is added and then the project is built then any
unit tests should then be displayed.

This fix is still not ideal since really the project should
not be displayed at all if there are no tests. So this approach may
need to be revisited.

Fixes #35